test: split upload-and-write.test.ts into focused per-feature files
The previous single file name implied upload + write coverage, but the
file actually held tests for ~10 separate concerns (download, raw, path
validation, shutdown, volume discovery, etc). Split into focused files:

- error-handling.test.ts   _handleApiError + _sendStatusError
- upload.test.ts            upload streaming + cache invalidation
- raw-endpoint.test.ts      /raw security headers
- download-endpoint.test.ts /download Content-Disposition
- delete.test.ts            delete + cache invalidation
- mkdir.test.ts             mkdir + cache invalidation
- shutdown.test.ts          shutdown + trackWrite
- volume-config.test.ts     discoverVolumes + clientConfig
- path-validation.test.ts   null-byte / 4096-cap / required-path

Shared pure helpers (mockReq/mockRes/getRouteHandler/etc) extracted to
_test-helpers.ts. The vi.hoisted/vi.mock block stays inlined per file
to match the convention in plugin.test.ts. No test logic changes:
50/50 still passing in the files plugin (1242 across appkit).
